LayoutTests/fast/webgpu/array-length-pointer-parameter-expected.txt

PASS output[0] is 13
PASS output[1] is 1
PASS output[2] is 0
PASS output[3] is 3
PASS output[4] is 0
